provide heading lvl 1 to app
the app had no heading level 1, but the SVG within the header could provide this information.

This PR modifies the header’s contents to wrap the svg within an h1 element.

The `h1` then has visually hidden text (the span w/at-only class) to announce the text of the SVG.

the `svg` itself is hidden from assistive technologies, and set to `focusable=“false”` so that IE won’t allow keyboard focus to enter it.

(safari + VO has some issues with SVG announcments, otherwise I would have added the hidden text as an `aria-label` to the svg, and given it a `role=“img”`…)